pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "document-intelligence-refinery"
version = "0.1.0"
description = "Enterprise-grade document intelligence pipeline with multi-strategy extraction, semantic chunking, and provenance tracking"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Tsegay", email = "tsegayassefa27@gmail.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Topic :: Text Processing :: General",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pydantic>=2.12.5",
    "pdfplumber>=0.11.0",
    "pymupdf>=1.24.0",
    "pillow>=10.0.0",
    "python-magic>=0.4.27",
    "langgraph>=1.0.10",
    "langchain-core>=1.2.16",
    "click>=8.1.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.0",
    "tenacity>=9.0.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
]

[project.scripts]
refinery-triage = "src.cli:triage"
refinery-extract = "src.cli:extract"
refinery-chunk = "src.cli:chunk"
refinery-index = "src.cli:index"
refinery-query = "src.cli:query"
refinery-audit = "src.cli:audit"
refinery-demo = "src.cli:demo"

# ⭐ CRITICAL FIX: Add this section!
[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
    "/pyproject.toml",
    "/rubric",
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
pythonpath = ["."]  # This helps pytest find your modules!